pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wafer_predictor"
version = "0.0.1"
description = "Predictive maintenance for silicon wafers: using sensor-driven AI to preemptively identify defects and streamline semiconductor manufacturing"
authors = [
  { name = "PreritSM" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License"
]
requires-python = "~=3.12.0"
dependencies = [
    "boto3>=1.40.0",
    "dvc>=3.66.1",
    "dvc-gdrive>=3.0.1",
    "fastapi>=0.116.0",
    "httpx>=0.28.0",
    "matplotlib>=3.10.8",
    "mlflow>=1.27.0",
    "pandas>=3.0.0",
    "pydantic-settings>=2.10.1",
    "psycopg2-binary>=2.9.10",
    "pyyaml>=6.0.3",
    "ruff>=0.14.14",
    "scikit-learn>=1.8.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
    "xgboost>=3.1.3",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "isort",
    "flake8",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]


[tool.ruff]
line-length = 99
src = ["src"]
include = ["pyproject.toml", "src/**/*.py"]

[tool.ruff.lint]
extend-select = ["I"]  # Add import sorting

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-sort-within-sections = true

